Split db-discovery ANTLR g4 file and visitor (#10815)
* Split g4 file for db-discovery

* Split g4 file for db-discovery

* Split g4 file for db-discovery

* Fix code format
