+++ shardingsphere-features/shardingsphere-db-discovery/pom.xml
@@ -32,7 +32,8 @@
     <modules>
         <module>shardingsphere-db-discovery-api</module>
         <module>shardingsphere-db-discovery-core</module>
-        <module>shardingsphere-db-discovery-spring</module>
+        <module>shardingsphere-db-discovery-distsql</module>
         <module>shardingsphere-db-discovery-provider</module>
+        <module>shardingsphere-db-discovery-spring</module>
     </modules>
 </project>
