+++ shardingsphere-features/shardingsphere-db-discovery/shardingsphere-db-discovery-distsql/src/main/antlr4/imports/db-discovery/RQLStatement.g4
@@ -15,90 +15,14 @@
  * limitations under the License.
  */
 
-lexer grammar Keyword;
+grammar RQLStatement;
 
-import Alphabet;
+import Keyword, Literals, Symbol;
 
-WS
-    : [ \t\r\n] + ->skip
+showDatabaseDiscoveryRules
+    : SHOW DB_DISCOVERY RULES (FROM schemaName)?
     ;
 
-CREATE
-    : C R E A T E
-    ;
-
-ALTER
-    : A L T E R
-    ;
-
-DROP
-    : D R O P
-    ;
-
-SHOW
-    : S H O W
-    ;
-
-SHARDING
-    : S H A R D I N G
-    ;
-
-RULE
-    :  R U L E
-    ;
-
-FROM
-    : F R O M
-    ;
-
-RESOURCES
-    : R E S O U R C E S
-    ;
-
-GENERATED_KEY
-    : G E N E R A T E D UL_ K E Y
-    ;
-
-DEFAULT_TABLE_STRATEGY
-    : D E F A U L T UL_ T A B L E UL_ S T R A T E G Y
-    ;
-
-TABLE
-    : T A B L E
-    ;
-
-SHARDING_COLUMN
-    : S H A R D I N G UL_ C O L U M N
-    ;
-
-TYPE
-    : T Y P E
-    ;
-
-NAME
-    : N A M E
-    ;
-
-PROPERTIES
-    : P R O P E R T I E S
-    ;
-
-COLUMN
-    : C O L U M N
-    ;
-
-BINDING
-    : B I N D I N G
-    ;
-
-BROADCAST
-    : B R O A D C A S T
-    ;
-
-RULES
-    : R U L E S
-    ;
-
-COLUMNS
-    : C O L U M N S
+schemaName
+    : IDENTIFIER
     ;
