+++ shardingsphere-features/shardingsphere-db-discovery/shardingsphere-db-discovery-distsql/src/main/java/org/apache/shardingsphere/dbdiscovery/distsql/parser/core/DatabaseDiscoveryRuleLexer.java
@@ -15,66 +15,18 @@
  * limitations under the License.
  */
 
-lexer grammar Keyword;
+package org.apache.shardingsphere.dbdiscovery.distsql.parser.core;
 
-import Alphabet;
+import org.antlr.v4.runtime.CharStream;
+import org.apache.shardingsphere.distsql.parser.autogen.DatabaseDiscoveryRuleStatementLexer;
+import org.apache.shardingsphere.sql.parser.api.parser.SQLLexer;
 
-WS
-    : [ \t\r\n] + ->skip
-    ;
-
-CREATE
-    : C R E A T E
-    ;
-
-ALTER
-    : A L T E R
-    ;
-
-DROP
-    : D R O P
-    ;
-
-SHOW
-    : S H O W
-    ;
-
-RULE
-    :  R U L E
-    ;
-
-FROM
-    : F R O M
-    ;
-
-READWRITE_SPLITTING
-    : R E A D W R I T E UL_ S P L I T T I N G
-    ;
-
-WRITE_RESOURCE
-    : W R I T E UL_ R E S O U R C E
-    ;
-
-READ_RESOURCES
-    : R E A D UL_ R E S O U R C E S
-    ;
-
-AUTO_AWARE_RESOURCE
-    : A U T O UL_ A W A R E UL_ R E S O U R C E
-    ;
-
-TYPE
-    : T Y P E
-    ;
-
-NAME
-    : N A M E
-    ;
-
-PROPERTIES
-    : P R O P E R T I E S
-    ;
-
-RULES
-    : R U L E S
-    ;
+/**
+ * SQL lexer for database discovery rule.
+ */
+public final class DatabaseDiscoveryRuleLexer extends DatabaseDiscoveryRuleStatementLexer implements SQLLexer {
+    
+    public DatabaseDiscoveryRuleLexer(final CharStream input) {
+        super(input);
+    }
+}
